requirements.txt
# Python dependencies for SAR-to-EO Image Translation using CycleGAN + CBAM
# Install via:  pip install -r requirements.txt

# -----------------------------------------------------------------------
# Deep learning core
# -----------------------------------------------------------------------
torch>=2.0.0
torchvision>=0.15.0

# -----------------------------------------------------------------------
# Satellite imagery I/O
# -----------------------------------------------------------------------
rasterio>=1.3.0
numpy>=1.24.0
Pillow>=9.5.0

# -----------------------------------------------------------------------
# Loss functions and metrics
# -----------------------------------------------------------------------
pytorch-msssim>=1.0.0

# -----------------------------------------------------------------------
# Visualization
# -----------------------------------------------------------------------
matplotlib>=3.7.0

# -----------------------------------------------------------------------
# Training utilities
# -----------------------------------------------------------------------
tqdm>=4.65.0

# -----------------------------------------------------------------------
# Configuration
# -----------------------------------------------------------------------
PyYAML>=6.0
